Refactored out of main
