+++ TAAM/app/src/main/java/com/example/taam/TaamItem.java
@@ -1,10 +1,14 @@
 package com.example.taam;
 
-public abstract class TaamItem {
+public class TaamItem {
     int lotNumber;
     String name;
     String category;
     String period;
     String description;
     //missing field for file
+
+    public TaamItem() {
+
+    }
 }
